+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- GetRandom Extension Function for `Vector2Int` and `Vector2`
 - MinMaxLimit attribute
 
 ## [3.1.0] - 2025-06-22